fixup compilation error with ARM asm by not using integrated assembler, fixup compiler warnings from xcode 4.5

+++ Classes/AVAnimator/AVAnimatorMedia.m
@@ -655,7 +655,7 @@
 	OSStatus result =
 	AudioSessionSetProperty(kAudioSessionProperty_AudioCategory, sizeof(sessionCategory), &sessionCategory);
 	if (result != 0) {
-		NSLog(@"%@", [NSString stringWithFormat:@"AudioSessionSetProperty(kAudioSessionProperty_AudioCategory,kAudioSessionCategory_MediaPlayback) error : %d", result]);
+		NSLog(@"%@", [NSString stringWithFormat:@"AudioSessionSetProperty(kAudioSessionProperty_AudioCategory,kAudioSessionCategory_MediaPlayback) error : %d", (int)result]);
 	}
 }
